Post React Router cleanup
