+++ humanlayer-wui/src/AppStore.ts
@@ -1,0 +1,49 @@
+import type { SessionInfo } from '@/lib/daemon/types'
+import { create } from 'zustand'
+
+interface StoreState {
+  /* Sessions */
+  sessions: SessionInfo[]
+  focusedSession: SessionInfo | null
+  initSessions: (sessions: SessionInfo[]) => void
+  setFocusedSession: (session: SessionInfo | null) => void
+  focusNextSession: () => void
+  focusPreviousSession: () => void
+}
+
+export const useStore = create<StoreState>(set => ({
+  sessions: [],
+  focusedSession: null,
+  initSessions: (sessions: SessionInfo[]) => set({ sessions }),
+  setFocusedSession: (session: SessionInfo | null) => set({ focusedSession: session }),
+  focusNextSession: () =>
+    set(state => {
+      const { sessions, focusedSession } = state
+      if (sessions.length === 0) return state
+
+      const currentIndex = focusedSession ? sessions.findIndex(s => s.id === focusedSession.id) : -1
+
+      // If no session is focused or we're at the last session, focus the first session
+      if (currentIndex === -1 || currentIndex === sessions.length - 1) {
+        return { focusedSession: sessions[0] }
+      }
+
+      // Focus the next session
+      return { focusedSession: sessions[currentIndex + 1] }
+    }),
+  focusPreviousSession: () =>
+    set(state => {
+      const { sessions, focusedSession } = state
+      if (sessions.length === 0) return state
+
+      const currentIndex = focusedSession ? sessions.findIndex(s => s.id === focusedSession.id) : -1
+
+      // If no session is focused or we're at the first session, focus the last session
+      if (currentIndex === -1 || currentIndex === 0) {
+        return { focusedSession: sessions[sessions.length - 1] }
+      }
+
+      // Focus the previous session
+      return { focusedSession: sessions[currentIndex - 1] }
+    }),
+}))
