+++ humanlayer-wui/src/components/internal/SessionTable.tsx
@@ -76,7 +76,7 @@ export default function SessionTable({
             className={`cursor-pointer ${focusedSession?.id === session.id ? '!bg-accent/20' : ''}`}
           >
             <TableCell>{session.status}</TableCell>
-            <TableCell>{session.query}</TableCell>
+            <TableCell className="max-w-xs truncate">{session.query}</TableCell>
             <TableCell>{session.model || <CircleOff className="w-4 h-4" />}</TableCell>
             <TableCell>{session.start_time}</TableCell>
             <TableCell>{session.last_activity_at}</TableCell>
